chore: set final number for VIP 184 and its description
